Bump log4j version to mitigate CVE-2021-44228
CVE-2021-44228 allows attackers to execute custom code on the server, e.g. when passing it via user agent string when it is logged. It should be mitigated ASAP by raising the log4j version used in Blynk Server. I have not tested it and don't know about other changes in this versions, so take this as heads up which should be tested before merging.

Signed-off-by: MichaIng <micha@dietpi.com>